backend_v2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
apscheduler==3.10.4
python-multipart==0.0.6
gunicorn==21.2.0